move playground code to new package main.swift
